explicit type casts

+++ ww/utils/md5.c
@@ -100,7 +100,7 @@ void wwMD5Encode(unsigned char *output,unsigned int *input,unsigned int len) {
 void wwMD5Decode(unsigned int *output,unsigned char *input,unsigned int len) {
     unsigned int i = 0,j = 0;
     while(j < len) {
-        output[i] = (input[j]) | (input[j+1] << 8) | (input[j+2] << 16) | (input[j+3] << 24);
+        output[i] = ((unsigned int)input[j]) | (((unsigned int)input[j+1]) << 8) | (((unsigned int)input[j+2]) << 16) | (((unsigned int)input[j+3]) << 24);
         i++;
         j+=4;
     }
